sql/delegate: move delegateShowDatabaseIndexes to its own file
Release note: None
