+++ pkg/sql/delegate/show_database_indexes.go
@@ -1,0 +1,37 @@
+// Copyright 2019 The Cockroach Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+// implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package delegate
+
+import (
+	"fmt"
+
+	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
+)
+
+func (d *delegator) delegateShowDatabaseIndexes(
+	n *tree.ShowDatabaseIndexes,
+) (tree.Statement, error) {
+	const getAllIndexesQuery = `
+    SELECT table_name,
+           index_name,
+           non_unique::BOOL,
+           seq_in_index,
+           column_name,
+           direction,
+           storing::BOOL,
+           implicit::BOOL
+      FROM %s.information_schema.statistics`
+	return parse(fmt.Sprintf(getAllIndexesQuery, n.Database.String()))
+}
