+++ pkg/sql/delegate/show_table.go
@@ -34,22 +34,6 @@ func (d *delegator) delegateShowCreate(n *tree.ShowCreate) (tree.Statement, erro
 	return d.showTableDetails(n.Name, showCreateQuery)
 }
 
-func (d *delegator) delegateShowDatabaseIndexes(
-	n *tree.ShowDatabaseIndexes,
-) (tree.Statement, error) {
-	const getAllIndexesQuery = `
-    SELECT table_name,
-           index_name,
-           non_unique::BOOL,
-           seq_in_index,
-           column_name,
-           direction,
-           storing::BOOL,
-           implicit::BOOL
-    FROM %s.information_schema.statistics`
-	return parse(fmt.Sprintf(getAllIndexesQuery, n.Database.String()))
-}
-
 func (d *delegator) delegateShowIndexes(n *tree.ShowIndexes) (tree.Statement, error) {
 	const getIndexesQuery = `
     SELECT table_name,
